raft: Detect remote address on Join
Previously, we required a joining node to provide its remotely-reachable
address in the Addr field of the Join request. This places a burden on
admins because they must set up every node that may potentially become a
manager with an IP address in mind.

Allow an address of the form "0.0.0.0:4242" in Addr, and if this is
found, replace it with the address seen by the manager. Add support for
forwarding remote addresses in GRPC headers, for cases where the join
request is proxied.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>