Support plaintext secrets
Our secret message contains an Alg field. Currently the only supported
value is bcrypt.

We want to move in the direction of generating random secrets instead of
allowing user-specified passwords. To do make the automatically
generated secrets more discoverable, add support for a "plaintext" Alg.

bcrypt is still supported in case someone wants to use the API or a
custom client to set a bcrypted hash, but the intent is for Docker to
use the plaintext Alg only, with randomly-generated content.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>