raft: Fix data race caused by ReplaceMemberConnection
ReplaceMemberConnection was overwriting field in a Member struct, and
this same member struct (not a copy) was being returned by GetMember. So
a different goroutine could potentially be using this struct while it's
being mutated.

Change ReplaceMemberConnection to make a copy of the struct and replace
the entry in the members map with the copy.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>