raft: Refuse message from incorrect member
A member could register into the raft and then change
its address by restarting from the same state but with
a different address.

This PR refuses a 'ProcessRaftMessage' if it comes from
a node with a correct ID but with a different address than
the one used at join time. This keeps the state machine safe
from this incorrect scenario as the address should be stable
for a raft member.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>