api: default log driver goes under task defaults
In general, we'd like to avoid pollute the top-level space of the
cluster configuration with specific items. This parameterizes task
creation using a `TaskDefaults` subsection. It is arguable that this is
part of the orchestrator, but it is possible to allow task creation
outside of the orchestrator in the future.

Signed-off-by: Stephen J Day <stephen.day@docker.com>