Track task/node allocation using ID
Currently the task and node allocations are being tracked by the ip
address allocated which is incorrect as it does not correctly specify
that this IP is assigned to this task. Rather as is done for services,
tasks and nodes allocation should be tracked by it's own ID in an
allocation set. If the ID is present in the set then it is
allocated. This avoids problems like double freeing the same IP which
might have been allocated to other objects and cause conflicts.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>