Allow prefix search on cluster, network, node, and task names
This fix tries to address the issue raised in docker/24270

where it was not possible to have a partial name match
when list services with name filter.

This fix is an extension. It updates swarmkit and
allows prefix search when name is provided as the filter
for listing clusters, networks, nodes, and tasks.

Related tests have been updated.

This fix is related to docker/24270.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>