controlapi: allow removing Manager node if not in raft memberlist
Currently if a node is in the process of joining as a Manager,
it is being issued a Manager cert and role. Although if this
node issues a 'docker swarm leave', it is impossible to remove
it as we check on the role as a failed precondition.

This PR allows to remove a node with 'docker node rm' if a
node has the manager role but is not yet accepted and
registered in the raft memberlist.

Fixes #1180

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>