Option for rolling updates not to proceed on failure
Add an enum to UpdateConfig that controls whether the update proceeds or
pauses on a failure. This will be extended to support additional
settings and modes in the future, such as failure thresholds and
automatic rollback.

Add UpdateStatus to the Service object. This contains the state that
keeps track of whether the update is running, finished, or paused. It
also contains a human-readable message about the update status.

Change the updater to pause the update if FailureAction is set to pause,
and a task started by the updater stops running or fails to start while
the update is happening.

Update swarmctl to support a new option to control the action on update
failure, and to print out the update state in "service inspect".

Add an updater test covering both modes.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>