raft: fix Campaign on restore if the node id is not in the progress list
When restarting a node from its state, logs are reapplied to
restore the cluster membership with `ConfChangeAddNode` log
entries.

There is a chance to reach the portion calling `Campaign` to
restart a single cluster member faster when we are still
in the process of restoring the memberlist. In this case
the node may not appear in the progress list and we might
call `MaybeUpdate` on the node id from the map without
sanity checking thus resulting in a panic.

To avoid that, check if the node is present in the Progress
list before trying to Campaign on node restore.

Signed-off-by: Alexandre Beslic <alexandre.beslic@gmail.com>