Fix flaky test TestUpdaterFailureAction
This test ran an update that encountered an error and checked that it
stopped after the first task. Because there was no update delay, the
updater could update more than one task before the failure was noticed
(the notification would race with the next update). Add an update delay
to prevent this from happening.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>