manager/scheduler: disable plugin filter
Since plugins are lazy loaded on use in the engine, tasks that require a
plugin that hasn't been used yet will never be dispatched until the
engine description reports it.

Rather than remove all this code, we have left it in place to be used
when plugin scheduling is added in the future.

Signed-off-by: Stephen J Day <stephen.day@docker.com>