Allow remote address autodetection when bound to a particular address
When a joining node had a particular listen address specified, but no
advertise address, the listen address was being leaked to the raft layer
instead of using remote autodetection. We should never leak the listen
address, because the advertise address how other hosts should reach the
new node.

If no advertise address is specified, change the address we pass down to
raft to 0.0.0.0:<port>. This triggers the autodetection.

This only applies to "swarm join", since an advertise address is
*always* passed to Manager.New for "swarm init" (it may be
autodetected).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>