ca: Don't close "started" channel until the CA server is actually ready
"started" is closed slightly too early, so tests could start sending
requests to the CA server before it was ready to handle them. This bug
was present for awhile, but it was hidden before the join-token changes
because context was also set too early, so isRunning would return true.

This only affects unit tests, since the "started" channel is just used
to make the tests wait for the server to start up. So there should be no
impact on real-world use.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>