manager: Fix start/stop race
m.connSelector is set in Run while the manager is starting up. If Stop
is called before this point, m.connSelector might be created after there
is any possibility of stopping it, which would result in a goroutine
leak.

Make Stop wait for the connSelector to be created, to avoid this
situation.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>