cmd: fix name prefix matching
Currently if there are two services, and one service is named
busybox while the other is busybox-top. It is not possible to
remove the busybox service by giving its fullname which is also
a common name prefix of both. The only way out is removing the
busybox-top service first. This looks not an acceptable behavior.

The same problem applies to clusters, nodes and networks.

Fix it by finding a full name match if prefix matches are more
than one.

Another way to fix it is that start a request with a full name
filter before doing it with a name prefix filter. But I think
this way just presents the same result but adds one more call
overhead to server.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>
v3: fail out if there are multiple full name matches.
v2: do not use node's Spec.Annotations.Name as defaut