Agent: fix agent quit when promotion failed
When promoting a node to manager failed, it should and could stay
working as agent role, although it's role is changed to manager
in advance. The ideal solution should change its role back to agent
as well, but it will get more parts involved and require a larger
change. It's a long term item. For now, keep the node running is
better than quiting. It's actually the same behavior as if the
m.Run fails.

Signed-off-by: Jin Xu <jinuxstyle@hotmail.com>
v2: quit only if the caller is waiting manager part ready