dispatcher: avoid creating timers when nothing has changed
Currently, the dispatcher Tasks loop wakes up every 50 ms to check if it
has modifications it needs to push out. This is an unnecessary wakeup on
a laptop running Docker, and potentially a lot of CPU use on a manager
servicing thousands of nodes.

This changes it to only create that timer when there are pending
modifications. It also is more efficient about reusing an existing timer
instead of creating new ones, so that 200 modifications don't result in
200 stale timers whose expiry will be ignored.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>