dispatcher: Always send task updates to agent when state <= ASSIGNED
The use of TaskEqualStable in the dispatcher to avoid sending
unnecessary task updates is problematic for global services. These tasks
only move to the ASSIGNED state once the scheduler confirms resources
are available. Since state changes are not considered relevant changes
and do not trigger assignment set updates, the dispatcher may never send
this update, and tasks can become stuck in the ASSIGNED state.

If a task is in the state ASSIGNED or below, always count a state change
as a modification, even if nothing else has changed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>