Make register block until the batch completes
This means that "swarm init" and "swarm join" won't return until the
information in the node object is complete.

Also, add selects with waits on <-d.ctx.Done() on writes to
processUpdatesTrigger, in case the dispatcher is shut down while
multiple RPC invocations are trying to write to this channel (very
unlikely).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>