raft: Don't forward or accept proposals
Only the leader should make proposals, so we can guarantee that no two
proposals will conflict. This commit prevents a follower from sending a
proposal, or the leader from accepting one. We already have code to
prevent this, but there are certain corner cases where a leader could
become a follower while it's proposing a value.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>