raft: ForceNewCluster flag should only apply at startup
ForceNewCluster tells raft to use the existing state to create a new
cluster with a single member. This means that when first reading the
WAL/snapshot, the other existing managers are removed.

However, it doesn't make sense to do this more than once. For example,
if other managers are added to the cluster, and one of them takes over
as the leader, and sends the initial leader a snapshot, that snapshot
should be accepted without modifications.

Change the ForceNewCluster flag to only apply on startup.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit aaa027372b6647a21b4fd2f3eac2c72290dc36b9)