ca: Fix flaky test, clean up test code
Several tests created certs that expire in 1 minute, and then waited up
to 2 seconds for them to be renewed. But the logic for renewing a cert
only renews it quickly if the cert is expiring in *less than* one
minute. Since CFSSL rounds NotBefore/NotAfter to the nearest minute, we
could end up with a cert that expires in slightly more than one minute,
and it would take some random fraction of that minute to get renewed.

Change the renewal logic to renew certs after a second if they expire
in <= 1 minute, not < 1 minute.

Clean up the related tests (remove an unnecessary for/select loop).

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>