raft: Make TestRaftSnapshotRestart more robust
Generally, restarting a follower node will not cause a leader election.
But our test advances the fake clock as part of its polling, and on a
slow machine this can make enough time pass from raft's point of view to
trigger an election.

Don't assume that node 1 will still be the leader after the restart.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>