raft: Avoid double lock of stopMu
There is a possible deadlock on shutdown because LeaderAddr acquires
stopMu for reading, then calls WaitForLeader, which calls Leader (which
also acquires stopMu for reading). When Shutdown is called between these
read locks, Stop is blocking waiting for the readers to finish, but
Leader is blocking, preventing LeaderAddr from releasing the lock.

The isShutdown check does not look necessary anyway. If the context is
cancelled, WaitForLeader will return early, and its calls to Leader are
protected by acquiring stopMu individually. Remove the isShutdown check
and the locking that was related to it.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>