Skip same service PortConfig check to allow PortConfig updating
This fix is related to
https://github.com/docker/docker/pull/25428
https://github.com/docker/swarmkit/issues/1348
https://github.com/docker/docker/issues/25375

Basically, when service is updated, a different PortConfig might
be provided. This should be treated as an update (should be successful).
Currently, this is treated as an error in swarmkit.

This fix skips the check when comparing the PortConfig from the request
and the stored value so that the service could be updated.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>
(cherry picked from commit aaf1efbf99e2c6404bab1ffeffb4f2bd2e3644d7)