orchestrator: Reconcile instances, not tasks
This commit prepares for an upcoming change that will allow a
configurable rollout strategy (either shut down the old task before
starting the new one, or start the new one before shutting down the old
one). The issue that was blocking having a rollout strategy with an
overlap between the lifetime of the new and old tasks was that
service-level reconciliation couldn't properly handle situations where
there was more than one active task per slot or node. This change
attempts to solve this by moving to *instance-level* reconciliation.

Before, the orchestrator counted tasks. With this change, it counts
instances (slots or nodes) that are satisfied with at least one task
that has a desired state > NEW and <= RUNNING.

When service-level reconciliation calls the updater, it passes
[][]*api.Task instead of a flat []*api.Task. The first dimension is the
instance (slot or node grouping), and the next dimension contains the
set of tasks for that instance. This allows the updater to handle
situations like partial updates. For example, if slot 5 has two running
tasks, and only one matches the current service spec, the updater can
recognize that this is a partial update, and remove the dirty task. In
this model, the orchestrator doesn't need to know anything about what's
going on with the individual tasks. It just has scale up and down as
appropriate, and let the updater reconcile state below the instance
level.

I explored a few alternative approaches as well, and ran into a lot of
complexity because the orchestrator invariably had to recognize partial
update situations, and take action to remediate them (making sure not to
fight with a current update in progress). I find this model more elegant
because the orchestrator only needs to call Update, as it did before,
and the updater handles all logic related to updates. But I'm interested
in feedback and I'm willing to explore more alternatives.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>