raft: Avoid closing the wrong connection on an error
ReplaceMemberConnection takes a new GRPC connection for a specific
member as an argument. It sets that member's connection to this new one,
and closes the existing GRPC connection.

The problem is that there may be a concurrent invocation of sendToMember
that's trying to send over the old connection. In this case, it will
also call ReplaceMemberConnection, which now closes the new connection
that it put in place, instead of the one where the error was actually
seen. If calls to sendToMember are likely to overlap, then this produces
a cascade of failures that are hard to recover from. Each invocation
closes the connection that another is using.

Fix this by having ReplaceMemberConnection only replace the connection
if the old connection is the one in the Member record. If it's not,
close the new connection instead. The new connection is still passed in
by the caller to avoid Dialing while holding the lock.

This appears to fix docker's TestApiSwarmLeaderElection on ARM.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit b5cc5bc4361910800354607b4e94d5b351a0d3ce)