Initial dispatcher support for secrets.
Assignments now provides a stream with incremental task and secret
updates. Additional object types can be supported in the assignment set
in the future. The first message returned from the Assignments stream is
the complete set of tasks and secrets for the node, and this is used to
synchronize the node's view with the manager's. Additional messages
returned by the stream are incremental updates that add, update, or
remove one or more tasks or secrets. If the agent gets out of sync with
the manager, it can reinitiate the Assignments stream to sync up.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>