Versioned specs
Add a SpecVersion field to Meta. Increment the spec version every time a
spec is updated through the control APIs.

Add a SpecVersion field to the Update* control API request messages.
This allows the update to apply to a particular spec version. Either the
object version or the spec version must be specified. If both are
specified, they are both enforced.

This is backward compatible because old clients won't send SpecVersion
in Update RPCs. Also, existing objects will be interpreted to have
SpecVersions of 0, and they will be incremented from there as soon as
the specs are updated with a new version of the manager.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>