dispatcher: Fix long-running TestSession unit test
This test calls Recv twice, but there doesn't seem to be a good reason.
The second call blocks for awhile until the node heartbeat expires
(because this test isn't sending heartbeats).

Presumably in the past the list of managers was only sent in the second
message, but now it's sent in the first one. So only call Recv once.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>