scheduler: implement HA scheduling
Change the scheduler to prioritize spreading out tasks from the same
service over spreading out the total number of tasks on each node.

This switches to an O(n) scheduling algorithm, but amortizes the cost of
evaluating all the nodes over a whole batch of tasks that share the same
spec, instead of doing it separately for each task.

This is an initial proof of concept and there are still many
opportunities for cleanup and optimization - see the TODO comments.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>