Change Task Name to `<ServiceAnnotations.Name>.<NodeID>.<TaskID>` in global mode
This fix is based on the discussion in:
https://github.com/docker/docker/pull/24850#issuecomment-243518250

Previously Task Name has been defined as `<ServiceAnnotations.Name>.<Slot>.<TaskID>`
in both replicated and global mode. That means in global mode, the Task Name is
`<ServiceAnnotations.Name>.0.<TaskID>` which does not carry any information in `0`.

In this fix, the Task Name has been changed to `<ServiceAnnotations.Name>.<NodeID>.<TaskID>`
in global mode.

This fix is related to:
https://github.com/docker/docker/pull/24850

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>