raft: Migrate WAL files to new directory
The currently vendored version of etcd/wal produces files that are
incompatible with the version of this package that 1.12 vendored. When
reading files created by this version, the WAL decoder in 1.12 will
panic.

To avoid causing crashes in downgrade scenarios, copy existing WAL files
to a new path that Docker 1.12 doesn't read from. The old files will be
preserved in case there is a downgrade. The downgraded version would
read the old state and catch up through raft.

We also migrate the snapshots (using hard links), so that after a
downgrade the snapshots produced by 1.12 from different WALs won't be
visible to > 1.12.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>