watch: Drop ErrSinkClosed return values from Write
Sometimes, if an event is being written during the moment when a sink is
removed from the broadcaster, it triggers a long and scary log message
saying that an event was dropped. In the use of go-events in this watch
package, it's expected for this event to be dropped, because we are
closing down the sink and not intending to receive further events on it.
Add a wrapper around the channel that makes Write succeed even if
ErrSinkClosed is returned.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>