cmd/swarmctl: implement simple service logs
For use in testing the logs feature, define a `swarmctl service logs`
command. It only only opens up a full stream to get all logs of all
services for now but we'll need to this base to test out filtering and
routing in the future.

In practice, this component is useless without the broker
implementation, which will follow later.

Signed-off-by: Stephen J Day <stephen.day@docker.com>