raft: Handle IP address changes
Record the most recently seen source address for each peer. If sending
to a member fails and it connected from a different address, try a
health check against that source address. If the health check passes,
switch the outbound connection to use that address. On the leader, write
a node update to raft that records the new address.

Block vote requests from unreachable members, so spurious elections
don't disrupt the cluster.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>