raft: Deduplicate node-level options
The Node structure has too many fields. One small step in the right
direction is removing fields that are directly copied from the
NewNodeOptions struct, which is also referenced in Node. There is less
potential for error if there is a single source of truth for these
options rather than storing the values in two places.

Rename NewNodeOpts to NodeOpts, since it's really used for more than
NewNode.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>