raft: Fix campaign on restart
In #1221, the code was changed not to call Campaign on restart unless
n.Node.Status().Progress[n.Config.ID] was set. The problem is that this
is never set unless the node is already the leader, so this effectively
disabled the quick leader election on restart for a single-node cluster.

Instead of checking Progress, figure out whether it's safe to call
Campaign by checking if we are in the member list. This list is kept in
sync with the config changes we apply in raft.

Also, only enable the calls to Campaign when the node was restarted from
state on disk.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>