Remove tasks that no longer satisfy constraints
If a node is in a way that makes its tasks no longer satisfy constraints
or resource requirements, shut down the affected tasks.

This is not done in the scheduler because the scheduler should not be
shutting down tasks. DesiredState is a field owned by the orchestrator,
and should not be written by other components.

Instead, introduce a ConstraintEnforcer inside the orchestrator package.
It waits for node updates, and checks that the tasks on the node still
satisfy constraints and resource requirements.

Change the scheduler to atomically check that nodes where tasks are
assigned haven't changed during the time the scheduling decision was
made, since this could cause a new task to be assigned to node that was
just updated in a way that makes it incompatible with the task.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>